Moved files into folders to decluter the repo
